move pagination logic to separate module
